[FEAT]오피스텔 상세 페이지[ADD]아파트 상세 페이지 생성
